feat(baselines): add hcnn.baselines with RNN/LSTM trainers
Group the RNN/LSTM comparison models under an explicit hcnn.baselines namespace
so their role (comparison, NOT part of the HCNN method) is unambiguous:

- move classic_models.py / classic_ensembles.py into hcnn/baselines/; leave
  deprecated re-export shims at the old hcnn.core paths so existing imports work.
- add SequenceModelTrainer (next-step prediction + autoregressive forecast) and
  SequenceEnsembleTrainer, both thin subclasses of the shared BaseTrainer /
  EnsembleTrainer, so baselines inherit every trainer fix without duplication.
- expose hcnn.baselines and the shared trainers from the top-level package.

Enables like-for-like HCNN-vs-baseline experiments on one data pipeline and one
training stack.
